requirements.txt
# requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# Google Cloud and Vertex AI
google-cloud-aiplatform
vertexai

# LangChain for RAG and Vector Search
langchain
langchain-core
langchain-google-vertexai
langchain-community

# LlamaIndex for Vector Store
llama-index
llama-index-vector-stores-vertexaivectorsearch
llama-index-embeddings-huggingface

# PDF Processing
PyMuPDF

# Machine Learning
transformers
torch
sentence-transformers
numpy

# Utilities
python-dotenv
aiofiles

# Monitoring and Logging (optional but recommended)
prometheus-client
python-json-logger

# Testing (optional, for development)
pytest
pytest-asyncio
httpx